\name{Say_greetings}
\alias{Say_greetings}

\title{
Say Christmas Greetings
}
\description{
This function says Christmas greetings to the name parameter
}
\usage{
Say_greetings(name = name)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{name}{
A String that represents the name of the person who is to be greated
}
}
\author{
Yogananda Srihari Prasad Gone
}
